Add support for filtering custom objects in where()
This replaces the function isPlainObject with isPrimitiveWrapper,
which checks if the given argument is a wrapper for one of the
primitive types string, boolean, number or BigInt. For this to work, the
order of some conditional branches had to be changed.

This now allows custom objects to be filtered in where().

Closes #14